Normalize path everywhere
Arrow is going to call normalize_path before sending us any path
input, so this should normally not change anything, and it would be
enough that we provide the functionality. By ensuring that we
normalize path everywhere, we ensure that internally, there's only one
valid string to represent each possible path.

We attempt to ensure that all leading and trailing slashes are removed
before sending anything to Azure. This may mean that you calling code
operates on files having a different path than they assume, but as
long as all access using that path happens through pyarrowfs-adlgen2,
we're all good.