Add optional metadata argument to output streams
This is an API upgrade in pyarrow 5 that we can make use of, that allows us to
easily set content-type and other metadata when opening files. Verified that
adding it this way does not break earlier versions of pyarrow for us. Document
where to find the valid keys for this where I could think of it.